Test 1 (Failed)

+++ Assets/Script/Scriptable Objects/TurretData.cs
@@ -10,10 +10,9 @@ public class TurretData : ScriptableObject
     public float fireRate;
     public float fireCountdown;
     public float range;
-    public int attackDamage;
+    public float attackDamage;
     public float rotationSpeed;
     public float slowAmount;
-
 
 
     public List<TurretData> levelTurret = new();
